% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_calc.R
\name{calculate_bic}
\alias{calculate_bic}
\title{Calculate BIC for Decomposition Results (Multi-View)}
\usage{
calculate_bic(Y, res)
}
\arguments{
\item{Y}{A list of original input matrices, each of size N × p.}

\item{res}{The output list returned by \code{multi_view_decomposition()}.}
}
\value{
A numeric vector of BIC values, one for each view.
}
\description{
Calculate BIC for Decomposition Results (Multi-View)
}
