% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{Ltrans}
\alias{Ltrans}
\title{Extract Upper Triangular Part of a Matrix}
\usage{
Ltrans(X, d = FALSE)
}
\arguments{
\item{X}{A square matrix.}

\item{d}{Logical, if \code{TRUE}, includes diagonal elements.}
}
\value{
A numeric vector of the upper triangular part of \code{X}.
}
\description{
Extract Upper Triangular Part of a Matrix
}
